+++ src/main/java/net/MrGise/mmm/screen/thingamajig/ThingamajigScreen.java
@@ -1,7 +1,9 @@
 package net.MrGise.mmm.screen.thingamajig;
 
 import com.mojang.blaze3d.systems.RenderSystem;
+import net.MrGise.floating.helper.MouseUtil;
 import net.MrGise.mmm.MMM;
+import net.MrGise.mmm.screen.renderer.EnergyDisplayTooltipArea;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.GameRenderer;
@@ -9,10 +11,12 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 
+import java.util.Optional;
+
 public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> {
     private static final ResourceLocation TEXTURE =
             new ResourceLocation(MMM.MOD_ID, "textures/gui/thingamajig/thingamajig_gui.png");
-
+    private EnergyDisplayTooltipArea energyDisplayArea;
 
     public ThingamajigScreen(ThingamajigMenu menu, Inventory inv, Component title) {
         super(menu, inv, title);
@@ -24,6 +28,24 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
         super.init();
         this.titleLabelY = 4;
         this.inventoryLabelY = this.imageHeight - 92;
+
+        energyDisplayArea = new EnergyDisplayTooltipArea(((width - imageWidth) / 2) + 156,
+                ((height - imageHeight) / 2) + 17, 8, 64, menu.blockEntity.getEnergyStorage());
+    }
+
+    @Override
+    protected void renderLabels(GuiGraphics gui, int mouseX, int mouseY) {
+        int x = (width - imageWidth) / 2;
+        int y = (height - imageHeight) / 2;
+
+        renderEnergyTooltip(gui, mouseX, mouseY, x, y);
+    }
+
+    private void renderEnergyTooltip(GuiGraphics gui, int mouseX, int mouseY, int x, int y) {
+        if (isMouseAboveArea(mouseX, mouseY, x, y, 156, 17, 8, 64)) {
+            gui.renderTooltip(this.font, energyDisplayArea.getTooltips(),
+                    Optional.empty(), mouseX - x, mouseY - y);
+        }
     }
 
     @Override
@@ -37,6 +59,8 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
         gui.blit(TEXTURE, x, y, 0, 0, imageWidth, imageHeight);
 
         renderProgressArrow(gui, x, y);
+
+        energyDisplayArea.render(gui);
     }
 
     private void renderProgressArrow(GuiGraphics gui, int x, int y) {
@@ -50,5 +74,9 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
         renderBackground(gui);
         super.render(gui, mouseX, mouseY, delta);
         renderTooltip(gui, mouseX, mouseY);
+    }
+
+    private boolean isMouseAboveArea(int mouseX, int mouseY, int x, int y, int offsetX, int offsetY, int width, int height) {
+        return MouseUtil.isMouseOver(mouseX, mouseY, x + offsetX, y + offsetY, width, height);
     }
 }